-Fix xspriData2Array function -Fix kModernSpriteDamager type
