+++ source/blood/src/gameutil.cpp
@@ -23,6 +23,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <random>
 #include "build.h"
 #include "common_game.h"
 
@@ -909,4 +910,23 @@ int picWidth(short nPic, short repeat) {
 
 int picHeight(short nPic, short repeat) {
     return ClipLow((tilesiz[nPic].y * repeat) << 2, 0);
+}
+
+bool xspriData2Array(int nXSprite, int* rData) {
+    if (xspriRangeIsFine(nXSprite) && rData) {
+        rData[0] = xsprite[nXSprite].data1; rData[2] = xsprite[nXSprite].data3;
+        rData[1] = xsprite[nXSprite].data2; rData[3] = xsprite[nXSprite].data4;
+        return true;
+    }
+
+    return false;
+}
+
+// by NoOne: used for better randomness in single player
+int STD_Random(int a, int b) {
+
+    std::default_random_engine rng;
+    rng.seed(std::random_device()());
+    std::uniform_int_distribution<int> dist_a_b(a, b);
+    return dist_a_b(rng);
 }
