+++ source/blood/src/gameutil.h
@@ -49,7 +49,7 @@ enum {
 };
 
 
-// by NoOne: functions to quckly check range of specifical arrays
+// by NoOne: functions to quickly check range of specifical arrays
 inline bool spriRangeIsFine(int nIndex) {
     return (nIndex >= 0 && nIndex < kMaxSprites);
 }
@@ -83,3 +83,5 @@ int GetClosestSectors(int nSector, int x, int y, int nDist, short *pSectors, cha
 int GetClosestSpriteSectors(int nSector, int x, int y, int nDist, short *pSectors, char *pSectBit, short *a8);
 int picWidth(short nPic, short repeat);
 int picHeight(short nPic, short repeat);
+bool xspriData2Array(int nXSprite, int* rData);
+int STD_Random(int a, int b);
