+++ source/blood/src/triggers.h
@@ -85,7 +85,4 @@ bool valueIsBetween(int val, int min, int max);
 void trPlayerCtrlLink(XSPRITE* pXSource, PLAYER* pPlayer);
 void trPlayerCtrlStartScene(XSPRITE* pXSource, PLAYER* pPlayer, int causedBy);
 void trPlayerCtrlStopScene(XSPRITE* pXSource, PLAYER* pPlayer);
-
-extern int gRdata[4];
-int* xspriData2Array(int nXSprite);
 // -------------------------------------------------------
